Moved code around to sort dependencies and functions visibility
